+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Fix FSD animation not stopping when entering supercruise
 - Fix lighting of some bindings that only apply on/off supercruise
+- Ignore duplicate EDHM configuration entries
 
 ## [1.16.0](https://github.com/poveden/EliteChroma/compare/v1.15.0...v1.16.0) — 2022-01-16